requirements_dev.txt
# core
typer[all]==0.4.1

# dev
black==22.3.0
commitizen==2.27.1
flake8==4.0.1
invoke==1.7.1
isort==5.10.1
pre-commit==2.19.0;python_version>='3.8'
pre-commit==2.17.0;python_version<'3.8'

# test
pytest-cov==3.0.0
pytest-logger==0.5.1
pytest-mock==3.7.0;python_version>='3.8'
pytest-mock==3.6.1;python_version<'3.8'
pytest-sugar==0.9.4
tox==3.25.0;python_version>='3.8'
tox==3.24.5;python_version<'3.8'

# docs
mkdocs==1.3.0;python_version>='3.8'
mkdocs==1.2.4;python_version<'3.8'
mkdocs_material==8.3.5;python_version>='3.8'
mkdocs_material==8.2.7;python_version<'3.8'

# dist
flit==3.7.1
